+++ elmyra.ip.access.epo/elmyra/ip/access/epo/services.py
@@ -56,8 +56,8 @@ def ops_published_data_search_handler(request):
     query = request.params.get('query', '')
     log.info('query raw: ' + query)
 
-    # fixup query: wrap into quotes if cql string contains spaces and is still unquoted
-    if ' ' in query and query[0] != '"' and query[-1] != '"':
+    # fixup query: wrap into quotes if cql string is a) unspecific, b) contains spaces and c) is still unquoted
+    if '=' not in query and ' ' in query and query[0] != '"' and query[-1] != '"':
         query = '"%s"' % query
 
     # Parse and recompile CQL query string to apply number normalization