+++ elmyra.ip.access.epo/CHANGES.rst
@@ -6,6 +6,7 @@ development
 ===========
 - ui: click on document-number in header to navigate to this document
 - ui: enhance pager, display active pagination entry, display current range
+- ui: open drill-down links in external window
 
 0.4.0
 =====